Revert "add menu-list"
This reverts commit 5799df9bf202eb17f7b784be7eb79404fce68e8f.
