+++ packages/menu/src/submenu.vue
@@ -3,7 +3,6 @@
   import menuMixin from './menu-mixin';
   import Emitter from 'element-ui/src/mixins/emitter';
   import Popper from 'element-ui/src/utils/vue-popper';
-  import MenuList from './menu-list';
 
   export default {
     name: 'ElSubmenu',
@@ -12,9 +11,13 @@
 
     mixins: [menuMixin, Emitter, Popper],
 
-    components: { ElCollapseTransition, MenuList },
+    components: { ElCollapseTransition },
 
     props: {
+      transformOrigin: {
+        type: [Boolean, String],
+        default: false
+      },
       index: {
         type: String,
         required: true
@@ -102,12 +105,17 @@
             ? this.activeTextColor
             : this.textColor
         };
-      },
-      popperPlacement() {
-        return this.mode === 'horizontal' ? 'bottom-start' : 'right-start';
       }
     },
     methods: {
+      handleCollapseToggle(value) {
+        if (value) {
+          this.initPopper();
+        } else {
+          this.doDestroy();
+          document.body.removeChild(this.popperElm);
+        }
+      },
       addItem(item) {
         this.$set(this.items, item.index, item);
       },
@@ -165,11 +173,23 @@
         if (this.mode === 'horizontal' && !this.rootMenu.backgroundColor) return;
         const title = this.$refs['submenu-title'];
         title && (title.style.backgroundColor = this.rootMenu.backgroundColor || '');
+      },
+      updatePlacement() {
+        this.currentPlacement = this.mode === 'horizontal' ? 'bottom-start' : 'right-start';
+      },
+      initPopper() {
+        this.referenceElm = this.$el;
+        this.popperElm = this.$refs.menu;
+        this.updatePlacement();
       }
     },
     created() {
       this.parentMenu.addSubmenu(this);
       this.rootMenu.addSubmenu(this);
+      this.$on('toggle-collapse', this.handleCollapseToggle);
+    },
+    mounted() {
+      this.initPopper();
     },
     beforeDestroy() {
       this.parentMenu.removeSubmenu(this);
@@ -184,9 +204,41 @@
         backgroundColor,
         $slots,
         rootMenu,
-        isMenuPopup,
+        currentPlacement,
+        menuTransitionName,
         mode
       } = this;
+
+      const popupMenu = (
+        <transition name={menuTransitionName}>
+          <div
+            ref="menu"
+            v-show={opened}
+            class={[`el-menu--${mode}`]}
+            on-mouseenter={this.handleMouseenter}
+            on-mouseleave={this.handleMouseleave}
+            on-focus={this.handleMouseenter}>
+            <ul
+              role="menu"
+              class={['el-menu el-menu--popup', `el-menu--popup-${currentPlacement}`]}
+              style={{ backgroundColor: rootMenu.backgroundColor || '' }}>
+              {$slots.default}
+            </ul>
+          </div>
+        </transition>
+      );
+
+      const inlineMenu = (
+        <el-collapse-transition>
+          <ul
+            role="menu"
+            class="el-menu el-menu--inline"
+            v-show={opened}
+            style={{ backgroundColor: rootMenu.backgroundColor || '' }}>
+            {$slots.default}
+          </ul>
+        </el-collapse-transition>
+      );
 
       return (
         <li
@@ -218,15 +270,7 @@
             }}>
             </i>
           </div>
-          <MenuList
-            visible={opened}
-            menu-popup={isMenuPopup}
-            append-to-body={isMenuPopup && this.$parent === rootMenu}
-            placement={this.popperPlacement}
-            on-mouseenter={this.handleMouseenter}
-            on-mouseleave={this.handleMouseleave}>
-            {$slots.default}
-          </MenuList>
+          {this.isMenuPopup ? popupMenu : inlineMenu}
         </li>
       );
     }
