+++ packages/theme-chalk/src/menu.scss
@@ -103,9 +103,6 @@
       border-bottom: 2px solid $--color-primary;
       color: $--color-text-primary;
     }
-    & .el-menu--popup {
-      padding: 5px 0;
-    }
   }
   @include m(collapse) {
     width: 64px;
@@ -162,9 +159,10 @@
     z-index: 100;
     min-width: 200px;
     border: none;
+    padding: 5px 0;
     border-radius: $--border-radius-small;
     box-shadow: $--box-shadow-light;
-    // overflow: hidden;
+    overflow: hidden;
 
     &-bottom-start {
       margin-top: 5px;
